requirements.txt
flask==2.3.2
flask-cors==4.0.0
gunicorn==21.2.0
sentence-transformers==2.2.2
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
huggingface-hub==0.16.4
python-dotenv==1.0.0
